django/requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cleanup==9.0.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
Markdown==3.7
sqlparse==0.5.3
